requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.3
scikit-learn==1.3.0
scikit-image==0.21.0
torch==2.1.0+cu118
torchvision==0.16.0+cu118
pandas==2.0.3
kaggle==1.5.16
tqdm==4.66.1
pyyaml==6.0.1
pillow==10.0.1
kagglehub==0.2.5

